pyproject.toml
[project]
name = "repl-nix-uibuilderpro"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.55.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "flask-login>=0.6.3",
    "flask-cors>=5.0.0",
    "pillow>=11.0.0",
]
